+++ sarscov2/model_mixins/__init__.py
@@ -1,1 +1,1 @@
-from .corona_kap_model_mixin import CoronaKapModelMixin
+from .corona_kap_model_mixin import CoronaKapModelMixin, CoronaKapDiseaseModelMixin