+++ CHANGES.rst
@@ -1,6 +1,16 @@
 Changes
 =======
 
+0.1.1
+-----
+- re-word economics questions on employment, health insurance, education (Okebe)
+- fix logic error for question of other sysptoms you know (Okebe)
+- remove precentage and change both questions to "Out of every 100 people, ..."
+- add questions about HIV, ART, diabetes and hypertension status
+- add indicator questions (weight, height, blood sugar)
+- split question on likelihood of infection -- 1) you, 2) your family
+- in knowledge, ask if more sever in those with diabetes, hypertension, HIV
+
 0.1.0
 -----
 - initial model, admin, form mixins and classes